refactor: move metadata rendering to macro
