+++ znap-syn/src/codegen/collection/handle_get.rs
@@ -1,6 +1,6 @@
 use proc_macro2::TokenStream;
 use quote::quote;
-use crate::{common::{create_get_context, create_get_handler, create_metadata, create_params}, CollectionMod};
+use crate::{common::{create_get_context, create_get_context_with_metadata, create_get_handler, create_metadata, create_params}, CollectionMod};
 
 pub fn generate(collection_mod: &CollectionMod) -> TokenStream {
     let impls: Vec<TokenStream> = collection_mod.actions
@@ -8,30 +8,43 @@ pub fn generate(collection_mod: &CollectionMod) -> TokenStream {
         .map(|action| {
             let handler = create_get_handler(&action.to_string());
             let context = create_get_context(&action.to_string());
+            let context_with_metadata = create_get_context_with_metadata(&action.to_string());
             let create_metadata_fn = create_metadata(&action.to_string());
-            
+            let params = create_params(&action.to_string());
+
             match &collection_mod.get_action_fns.iter().find(|get_action_fn| get_action_fn.action == *action) {
                 Some(get_action_fn) => {
                     let fn_block = &get_action_fn.raw_method.block;
-                    let params = create_params(&action.to_string());
-
+                    
                     quote! {
                         #[derive(Debug, serde::Serialize, serde::Deserialize)]
                         pub struct #context {
                             params: #params,
                         }
         
-                        pub async fn #create_metadata_fn(ctx: #context) -> znap::Result<znap::ActionMetadata> {
+                        #[derive(Debug, serde::Serialize, serde::Deserialize)]
+                        pub struct #context_with_metadata {
+                            params: #params,
+                            metadata: znap::ActionMetadata,
+                        }
+                        
+                        pub async fn #create_metadata_fn(ctx: #context_with_metadata) -> znap::Result<znap::ActionMetadata> {
                             #fn_block
                         }
         
                         pub async fn #handler(
                             axum::extract::Path(params): axum::extract::Path<#params>,
                         ) -> znap::Result<axum::Json<znap::ActionMetadata>> {
+                            let raw_metadata = #action::to_metadata();
                             let context = #context {
                                 params,
                             };
-                            let metadata = #create_metadata_fn(context).await?;
+                            let rendered_metadata = znap::render_metadata(&raw_metadata, &context, false, None);
+                            let context_with_metadata = #context_with_metadata {
+                                params: context.params,
+                                metadata: rendered_metadata,
+                            };
+                            let metadata = #create_metadata_fn(context_with_metadata).await?;
         
                             Ok(axum::Json(metadata))
                         }
@@ -39,8 +52,21 @@ pub fn generate(collection_mod: &CollectionMod) -> TokenStream {
                 },
                 _ => {
                     quote! {
-                        pub async fn #handler() -> znap::Result<axum::Json<znap::ActionMetadata>> {
-                            Ok(axum::Json(#action::to_metadata()))
+                        #[derive(Debug, serde::Serialize, serde::Deserialize)]
+                        pub struct #context {
+                            params: #params,
+                        }
+
+                        pub async fn #handler(
+                            axum::extract::Path(params): axum::extract::Path<#params>,
+                        ) -> znap::Result<axum::Json<znap::ActionMetadata>> {
+                            let raw_metadata = #action::to_metadata();
+                            let context = #context {
+                                params,
+                            };
+                            let rendered_metadata = znap::render_metadata(&raw_metadata, &context, false, None);
+
+                            Ok(axum::Json(rendered_metadata))
                         }
                     }
                 }
