+++ znap/Cargo.toml
@@ -18,6 +18,7 @@ solana-sdk = { workspace = true }
 # crate specific
 base64 = "0.22.1"
 bincode = "1.3.3"
+handlebars = "5.1.2"
 
 # local
 znap-macros = { path = "../znap-macros", version = "0.1.21" }
