+++ znap/src/prelude.rs
@@ -4,6 +4,6 @@ pub use crate::colored;
 pub use crate::tower_http;
 pub use crate::znap_macros::{collection, Action, ErrorCode};
 pub use crate::{
-    add_action_identity_proof, Action, ActionLinks, ActionMetadata, ActionResponse,
-    ActionTransaction, Error, ErrorCode, LinkedAction, LinkedActionParameter, Result, ToMetadata,
+    Action, ActionLinks, ActionMetadata, ActionResponse, ActionTransaction, Error, ErrorCode,
+    LinkedAction, LinkedActionParameter, Result, ToMetadata,
 };
